Campus/requirements.txt
Django==1.11.15
Django==2.1
Django==2.1.15
djangorestframework==3.14.0
PyMySQL==1.0.3
pytz==2018.5
pytz==2021.3
Requests==2.30.0
